requirements.txt
Flask==2.3.3
Flask-JWT-Extended==5.3.0
mysql-connector-python==8.4.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0